Camera events and started some turn experiments
Camera now has events for when it moves and rotates so that we don't update in billboard every frame, only when the camera rotates. Also created a new system where I'm trying to make a simpler dummy turn thing so that I can do some more testing with movement. Also making it so that I might be able to help with our current turn controller. I made a new scene and in that scene made severe changes to the menu on the canvas. It had a z position of like 150 which is absolutely weird. Also the scales were all wonky, so I fixed a lot of those. Those fixes are in my new scene, Wyatt_TurnTest